Fixed make targets so rebuilding works
The rebuild was doing funky things and resulting in weird errors. The makefile
was reworked so that rebuilding works correctly, and so that rebuilding can
restart from different stages of the build process. New `rebuild-with-linux`
target was added which restarts the kernel build, and the `rebuild` target was
changed so that only the rootfs is restarted.

Fixes #20

Signed-off-by: Branko Vukelic <branko@outernet.is>