Make initramfs compression configurable
This patch adds a configuration option

    User-provided options
        Firmware configuration
            Initramfs compression method

This is a three-state choice whith LZ4, GZIP and XZ options. Default
compression is GZIP. The mkcpio.sh has been modified to use this option to
determine the correct compression command.

Modified hookargs makefile and args.sh module relay this option to the hook
scripts including mkinitcpio.

Signed-off-by: Branko Vukelic <branko@outernet.is>