Somewhat bootable Pi3 setup
Long-story short, 4.4.y branch would not boot on Pi 3. In the interest of time,
and due mostly to the lack of output (rainbow screen is so goddamn helpful), I
decided to abandon 4.4 and use 4.1 instead. This is somewhat painful as it
increases the projected RAM usage, so we definitely want to go back to 4.4 once
things are stable.

The kernel configuration was swapped for the configuration that was originally
used for the ORxPi2 kernel.

Another big pain that this patch finally resolves is using separately built
compressed cpio archives for initramfs. In hindsight, it is hard to tell
whether this ever worked in Buildroot with rootfs-based cpio archives, but not
it works with the patches in the outernet branch, even with separately
generated ones.

In accordance with the Buildroot patch on outernet branch, the hookargs.mk has
been modified to relay initramfs compression-related setting to the mkcpio.sh
script. mkcpio.sh itself has been slightly modified to use compression settings
similar to gen_initramfs_list.sh from the kernel tree (in particular, use of
the -l switch for lz4 should be significant; see manpage for details).

A new post-image script has been added that attaches the Raspberry Pi trailer
(see the script source for details on that it does). The kernel compression had
to be switched back to gzip, however, as lz4 does not seem to be compatible
with the trailer.

The build process has been briefly documented in the build.rst.

Signed-off-by: Branko Vukelic <branko@outernet.is>