Added check for userspace init executable
When trying to `switch_root`, it is possible to attempt a boot of a root
filesystem that does not have a proper executable at /sbin/init. Since
attempting a `switch_root` to such a rootfs is suicide, a check is added that
tests whether there is an exectuable at /sbin/init.

While this check is not perfect, and won't be able to detect a malformed init
executable (one that does something stupid and causes the kernel to panic), it
does trap at least one case, so it's better than nothing.

Signed-off-by: Branko Vukelic <branko@outernet.is>