Move devtmpfs mountpoint to target rootfs
When kernel boots it will not automatically mount devtmpfs if there is an
initramfs. The init script will mount it, but it did not move the mount point
into the target rootfs making some weird noises during boot.

This patch fixes the issue, and also restores the previously unused and broken
`setup_boot` function.

The documentation is also updated to reflect the changes.

Signed-off-by: Branko Vukelic <branko@outernet.is>