Add support for .rxos override files
This patch allows for files that end in .rxos to be used as override files by
packages. The override files are used in situations where buildroot installs
files into the root filesystem *after* the packages install them. To prevent
package files from being overwritten, we can use .rxos files that will later be
renamed to correct path by the post-build hook in this patch.

Signed-off-by: Branko Vukelic <branko@outernet.is>