Build SD card image
This patch adds a set of configuration options that enable building of SD card
images using dir2fat32. The configuration options are grouped under a
menuconfig item in User-provided options > Firmware Configuration.

The dir2fat32 wrapper script, mksdimage.sh, is used to prepare the image. It is
fed a file pointed to by the configuration, that contains a list of files that
should be in the image. This file can be found in rxos/configs/sdimage.txt
which maps files in the $BINARIES_DIR to paths on the SD card.

Building of SD card image is turned on in the updated defconfig.

Signed-off-by: Branko Vukelic <branko@outernet.is>