Add support for boot-time partitioning
This patch introduces changes at several levels, starting with the build
itself. Building on the concept of local packages introduced with network
configuration, initramfs buiding was also converted into a local package. The
files and scripts related to initramfs building are now part of the new
ramfsinit local package.

The init script has been thoroughly reworked and:

- it now supports dropping straight to shell with `shell` kernel command line
- it supports adding multiple initramfs list fragments into the final initramfs
  list that is used to generate the cpio archive
- concept of boot hooks was introduced (and documented)
- generating cpio archives has been split between post-build hook and target
  finalize hooks, so access to build variables has been somewhat simplified
- procfs is mounted during early init, and the userspace proper now has a new
  `inittab` file that omits mounting of procfs

A boot hook called 'selfpartition' was added in the form of the
`boothook-selfpartition` local package. This package depends on ext2fs tools
(mke2fs in particular) and bundles a subset of libraries and binaries necessary
to perform ext4 formatting in the early userspace. It also adds the hook script
itself to the early userspace so it is executed during init.

The selfpartition boot hook will slice the SD card into multiple logical
partitions contained within a single extended partition that takes up the space
left after burning the SD card image onto a blank SD card. The minimal
requirement is around 4GB, although this leaves less than half a GB to the
downloads dir, so an 8GB card is really the usable minimum.

Because all partitioning and formatting is performed in the early userspace,
the usrspace proper now has a fstab file that manages the mount points.

Signed-off-by: Branko Vukelic <branko@outernet.is>