Add the ability to use rootfs overlays
This patch adds the ability to load arbitrary rootfs overlays that contain
extension and overrides for the base rootfs image. The init script has been
modified to mount and add overlays to the final OverlayFS mount. It does this
by discovering the overlay images on the SD card, and ignores the unmountable
ones.

The mksdimage.sh script was modified such that arbitrary files dropped into the
`$BINARIES_DIR/sdcard-extra` can be added to the SD card image. The intention
is to use this for including overlays in the build.

Signed-off-by: Branko Vukelic <branko@outernet.is>