Persist system configuration and state
This patch modifies the build to support persisten system configuration. It
introduces the S00 init script which allows arbitrary setup scripts to run
during init, and adds a persist setup script which performs configuration
persistence.

The files that should be persisted are listed in rxos/configs/persist.conf
which is copied into the rootfs at build time.

Additionally, a new dropbear init script is added which removes the unnecessary
fiddling with symlinks which is present in the original upstream script.

Signed-off-by: Branko Vukelic <branko@outernet.is>