Add application stack
In short, this commit adds the following parts of the application stack:

- Librarian
- FSAL
- monitoring daemon
- PostgreSQL
- lighttpd

Several local packages were added to provide kconfig interfaces for configuring
bits and pieces of the application stack:

- librarian-config
- fsal-config
- lighttpd-config
- postgres-config

A few of the other local pcakges were fixed in the process.

The pckages submodule has been updated to use the new packages with
configuation options (ondd, monitoring), and these now integrate with the
kconfig options provided in this build. A cleaner separation will be needed in
futrue relases so that packages don't end up with a hard dependency on
rxOS-style build.

The 'postgres' and 'data' users were added (clean build recommended if you have
a previous build) to allow PostgeSQL and ondd to configure their users.

Although it appears to be adding nginx, that is not the case: it's just a
package for installing nginx configuration which is not activated by default.

Due to locale issues, the entire toolchain has been downgraded to use glibc
2.21 instead of 2.22. Upstream patch for glibc was never tested.

A patch to the markers hook now causes the rootfs to contain `/etc/version` and
`/etc/platform` as in the other Outernet receivers.

The `collect_static.sh` script has been reinstated to perform the task of
collecting the static assets as in the ORx build.

Signed-off-by: Branko Vukelic <branko@outernet.is>