Fix ncurses compatibility with demod binary
Apparently, there was some left-over libncurses libary in the v1.0a1 build that
helped demod work on that image after a dirty rebuild. We switched to
libncurses with wide character support later and that removed the
libncurses.so.* files from the rootfs on the next clean rebuild, and broke the
demod.

This patch reverts to non-wide-char version of libnurses and fixes the libtinfo
symlink. We also moved the symlinking from a post-build script into a target
finalize hook, since using hooks as much as possible is preferred.

Signed-off-by: Branko Vukelic <branko@outernet.is>