Perform fsck on boot partition in early userspace
The FAT32 partition on the SD card is almost guaranteed to get corrupted when
power is yanked. This seems to happen regardless of whether there are writes
going to the partition or not. To handle the corruption, this commit adds
`fsck.vfat` to the initramfs image, and modifies the init script so that it
performs fsck on the boot partition, removes any .REC files that are created by
fsck, and remounts the partition read-only.

Signed-off-by: Branko Vukelic <branko@outernet.is>