Make /direct/ URL downloadable
Modern browsers usually do not try to download files they can render (e.g.,
video, audio, images, text), when pointed to them using a direct URL. To force
browsers to download, Content-Disposition header is needed. This patch changes
the lighty configuration such that the Content-Disposition header is added to
all /direct/ URLs. As a result, the download button in the UI has a more
natural behavior.

Signed-off-by: Branko Vukelic <branko@outernet.is>