Only add Content-Disposition to real download URLs
Because the direct URls are used in iframes, and other places, and not just for
the download link, the Content-Disposition header may interfere with normal
operation. This patch limits the addition of Content-Disposition to only the
/direct/ URLs that have a dl=1 query string. It is expected that the
application will make use of this to specify which links should have a
Content-Disposition header.

Signed-off-by: Branko Vukelic <branko@outernet.is>