Add local package for creating UBI image
For the multi-volume setup we need, we are using the standard CHIP MTD
partition layout and replacing the UBI image that is normally flashed on stock
CHIP by multi-volume UBI image. This commit adds a local package that contains
a post-image script and ubinize.cfg file for the multi-volume UBI image.
Enabling this package will use the built binaries and create the correct
multi-volume UBI image.

Signed-off-by: Branko Vukelic <branko@outernet.is>