Make Ethernet configuration configurable
CHIP does not have a (real) Ethernet port, so this commit makes the
network-config package a bit more flexible about enabling the support for it.
The new toggle option will cause the installation of dhcp/ifplugd/interfaces
related files. This allows us to turn it on for RPi3, and keep it off for CHIP.

Signed-off-by: Branko Vukelic <branko@outernet.is>