Add support for i2c-based LED control
Like Raspberry Pi, CHIP has a single user-controllable LED, but the i2c-tools
are used to control it, rather than sysfs. This commit modifies the led-control
package to provide i2c-based API for controlling the LED that is a drop-in
replacement for the old sysfs-based led_ctrl script.

The two versions can be toggled using a Kconfig option, and a special blink
script is provided for i2c version which allows simulation of timer trigger.

Signed-off-by: Branko Vukelic <branko@outernet.is>