Make LED signaling consistent
The hotplug script resets the LED to on after it completes. Due to the way i2c
API works, it is not possible to reset it to blinking mode, and therefore, it's
no feasible to use blinking as default LED state. On the other hand, LED starts
in constant-on state on power-on, so we need some form of indication that would
tell us that something changed (e.g., boot completed).

This patch changes the behavior of the S99led init script such that it blinks
the LED for 60 seconds before switching back to solid-on, therefore giving
enough time for the user to notice boot has completed, while maintaining
consistent LED signaling during hotplug events.

Signed-off-by: Branko Vukelic <branko@outernet.is>