Allow execution of pkg files on hotplug
This commit adds a handling of .pkg files on hotplug. This feature can be
toggled on and off using a new Kconfig option BR2_STORAGE_HOTPLUG_CHECK_PKG
housed under the storage-hotplug package's menu, which in turns toggles a flag
in the hotplug script. The hotplug script will always contain the code for
running .pkg files regardless of this option, and only the flag is toggled.

Signed-off-by: Branko Vukelic <branko@outernet.is>