Enable gadget serial
This patch changes the kernel configuration to include the gadget serial module
as built-in, and also adds a setup hook to modprobe it. The boothook is
converted to a setup hook to facilitate this. The inittab is also patched to
allow the gadget serial to be used as log-in shell.

Signed-off-by: Branko Vukelic <branko@outernet.is>