Make persistent config more dynamic
Using a flat file containing persistent files list does not yield well to
programmatic modification of that list by various packages. This commit
switches the persistent file list to a make variable so that it can be
overridden by packages. The network-config package is also updated to take
advantage of this feature.

Signed-off-by: Branko Vukelic <branko@outernet.is>