Add sdr-config package
This commit adds the sdr-config package which provides an init script and
script configuration for running the demod binary. Unlike standard init
scripts, it also provides a reset command which forcibly terminates the demod
binary and restarts it with 60 second interval. The 60-second pause is
syncrhonous so the init script blocks while waiting for restart.

Signed-off-by: Branko Vukelic <branko@outernet.is>