Move sdr init to 90 and fix timeout
The timeout value when resetting is required to be way longer than 60s. This
patch bumps it to 900s and also converts it to a variable at the top of the
init script for easier maintenance. The order of the script is also changed so
it runs _after_ ondd and not before.

Signed-off-by: Branko Vukelic <branko@outernet.is>