Turn debug flag on to improve performance
The debug flag controls the `filesystem_checks` parameter in Mako,
which causes a `stat` call on each template prior to rendering, to
check if the timestamp on the sources have changed since the last
time they were compiled. In case the timestamps differ the template
is recompiled. For our purposes it can be safely turned off, which
does provide a significant performance boost.