Run demod in screen to allow access to the debug UI
This patch changes the sdr init script such that the sdr binary is started in a
screen session rather than directly. This allows the debug curses UI to be
accessible while the demod is running. This also makes the sdr-config local
package to depend on screen.

A utility script, 'sdr', is provided to allow quick access to the debug UI.

Signed-off-by: Branko Vukelic <branko@outernet.is>