Add ability to read key password from pwd file
During signing of .pkg files, signing key password is required. Mistyping them
causes failed .pkg build and necessitates a rebuild. This patch adds the
ability to read the password from a password file named .password in the root
of the source tree instead of interactively. When this file is not present,
interactive password entry is still used.

Signed-off-by: Branko Vukelic <branko@outernet.is>