Use external's commit hash, not buildroot's
During the build, most of the targets are executed in the buildroot submodule's
directory. Because of this, the markers support makefile generates the commit
hash for the platform-release based on buildroot's HEAD instead of the whole
rxOS repository. This commit fixes that so that platform-release always points
to rxOS HEAD instead of buildroot.

Signed-off-by: Branko Vukelic <branko@outernet.is>