Better backup overlay handling
This commit modifies the init script to better handle overlay mounting. When an
unmountable overlay is encountered, the init will now attempt to mount the
backup copy (if any), and will lot a message about corrupted overlays if
backup cannot be mounted.

In the example messages below, 'foo' is a corrupted overlay that has no backup
copy, 'bar' is an overlay with both the master and backup copies corrupted, and
presets have a corrupted master but healthy master copy.

    Thu Jan  1 00:00:00 UTC 2015
    ++++ Starting rxOS v1.1 ++++
    ....
    Corrupted overlay bar
    Corrupted overlay foo
    Using overlay presets [BACKUP]

Closes #112

Signed-off-by: Branko Vukelic <branko@outernet.is>