Do not run hostapd as system service
Hostapd is currently running as a system service. While this is fine in most
cases, it makes it difficult for the end user to set up networking when AP mode
is not desired. This commit solves this issue by removing the init script and
replacing it with a new command, `ap` that is invoked as post-up and pre-down
hooks in the wifi interface configuration file.

The packages submodule is updated to a version that has configurable AP restart
command in librarian-netinterfaces, and an appropriate command is configured
for reloading the AP configuration.

Signed-off-by: Branko Vukelic <branko@outernet.is>