Add support for updating the U-Boot environment
This patch adds support for updating the U-Boot environment.

The chip-flash tool has been modified to output a copy of the environment that
is used in the script. This feature is meant to be used to syncrhonize the
rxos/configs/rxos_uboot-env file with the flashing script.

The installer has been modified to fix a bug that prevents U-Boot itself to be
flashed, and also adds support for flashing the U-Boot environment to the
appropriate MTD partition.

Finally, mkpkg is modified to pack the environment image, and also create a
separate .pkg file that contains only the images related to booting (U-Boot,
environment image, kernel image, and DTB).

Signed-off-by: Branko Vukelic <branko@outernet.is>