Added subplatform
This commit is a work-in-progress modification that adds a concept of
subplatforms to rxOS. Subplatform denotes a rxOS variant targeting a specific
board. Subplatforms may or may not be hardware-specific, and the intention is
for various OTA installers to use this marker to determine whether install is
required or not.

Originally, the idea was for CHIP and RPi3 installers to differentiate between
the respective boards by looking up the subplatform marker. However, it turns
out that the OTA update package does not even execute between these platforms
(triggering "Invalid instruction" errors), so subplatform is *not* used to
differentialte between CHIP and RPi3 boards in the OTA update installer script.

Signed-off-by: Branko Vukelic <branko@outernet.is>