Fix sdr init script
This commit modifies the demod init script to use screen directly instead of
using the start-stop-daemon wrapper. Since screen is a forking daemon, the
start-stop-daemon wrapper is not able to obtain its PID, and therefore is not
able to stop the session. In this patch, we use the screen's own mechanisms for
starting and stopping sessions.

Signed-off-by: Branko Vukelic <branko@outernet.is>